Change int cast to float

+++ exercises.py
@@ -7,5 +7,5 @@
 # Write a program to prompt the user for hours and rate per hour to compute gross pay
 hours = input('Enter Hours: ')
 rate = input('Enter Rate: ')
-pay = int(hours) * int(rate)
+pay = int(hours) * float(rate)
 print('Pay: ', pay)
